chore: update project

+++ EasyShop/app/src/main/java/pe/edu/upc/easyshop/core/root/Main.kt
@@ -20,10 +20,10 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.tooling.preview.Preview
 import pe.edu.upc.easyshop.core.ui.theme.EasyShopTheme
 import pe.edu.upc.easyshop.features.home.presentation.di.PresentationModule.getHomeViewModel
-import pe.edu.upc.easyshop.features.home.presentation.views.Home
+import pe.edu.upc.easyshop.features.home.presentation.home.Home
 
 @Composable
-fun Main(onClick: () -> Unit) {
+fun Main(onClick: (Int) -> Unit) {
     val selectedIndex = remember {
         mutableIntStateOf(0)
     }
